Finish verification of fib exaple programs
This includes the delay version and the state_delay version of fib.
Both are verified both as a top level definition using just the nth
number. But through helper lemmas that can start at any step, this is to
make the induction work.

The key to verifying both is the coq_fib_move' Lemma that let's us move
the addition out of the post condition from the iterative version into
the recursive version.

Remember to not simpl too much in wp_state_delay definitions since it
breaks the abstraction layer.