Progress on inductive case of fuel_adequacy
There seems to be a problem with how the modalities are iterated.
with the extra bupb being after the iterated modalities. Those need to
be eliminated before I can access information to use my induction
hypothesis. However the induction hypothesis will give me the iterated
modalities for a smaller n. This is a circle.