Start on bank example verification
The issue I face right now is how to let the value in the invariant vary
between acquisitions of the lock. I could quantify over the value
pointed to in the invariant, but then I lose the abillity to have the
n <= n' in the entire rule of withdraw_locked_swap.